move methods to ResourceUrlCreator
