+++ core/src/main/java/net/sourceforge/jnlp/cache/ResourceDownloader.java
@@ -1,18 +1,14 @@
 package net.sourceforge.jnlp.cache;
 
 import net.adoptopenjdk.icedteaweb.IcedTeaWebConstants;
-import net.adoptopenjdk.icedteaweb.client.parts.dialogs.security.InetSecurity511Panel;
-import net.adoptopenjdk.icedteaweb.client.parts.dialogs.security.SecurityDialogs;
 import net.adoptopenjdk.icedteaweb.commandline.CommandLineOptions;
 import net.adoptopenjdk.icedteaweb.http.CloseableConnection;
 import net.adoptopenjdk.icedteaweb.http.ConnectionFactory;
 import net.adoptopenjdk.icedteaweb.http.HttpMethod;
-import net.adoptopenjdk.icedteaweb.http.HttpUtils;
 import net.adoptopenjdk.icedteaweb.io.IOUtils;
 import net.adoptopenjdk.icedteaweb.jnlp.version.VersionString;
 import net.adoptopenjdk.icedteaweb.logging.Logger;
 import net.adoptopenjdk.icedteaweb.logging.LoggerFactory;
-import net.sourceforge.jnlp.DownloadOptions;
 import net.sourceforge.jnlp.runtime.Boot;
 import net.sourceforge.jnlp.runtime.JNLPRuntime;
 import net.sourceforge.jnlp.util.UrlUtils;
@@ -28,7 +24,6 @@ import java.io.OutputStream;
 import java.net.URL;
 import java.util.EnumSet;
 import java.util.HashMap;
-import java.util.List;
 import java.util.Map;
 import java.util.jar.JarOutputStream;
 import java.util.jar.Pack200;
@@ -47,61 +42,12 @@ public class ResourceDownloader implements Runnable {
 
     private static final Logger LOG = LoggerFactory.getLogger(ResourceDownloader.class);
 
-    private static final String ACCEPT_ENCODING = "Accept-Encoding";
-    private static final String PACK_200_OR_GZIP = "pack200-gzip, gzip";
-
-    private static final HttpMethod[] validRequestMethods = {HttpMethod.HEAD, HttpMethod.GET};
-
     private final Resource resource;
     private final Object lock;
 
     public ResourceDownloader(Resource resource, Object lock) {
         this.resource = resource;
         this.lock = lock;
-    }
-
-    /**
-     * Connects to the given URL, and grabs a response code and redirection if
-     * the URL uses the HTTP protocol, or returns an arbitrary valid HTTP
-     * response code.
-     *
-     * @return the response code if HTTP connection and redirection value, or
-     * HttpURLConnection.HTTP_OK and null if not.
-     * @throws IOException if an I/O exception occurs.
-     */
-    static UrlRequestResult getUrlResponseCodeWithRedirectionResult(final URL url, final Map<String, String> requestProperties, final HttpMethod requestMethod) throws IOException {
-
-        try (final CloseableConnection connection = ConnectionFactory.openConnection(url, requestMethod, requestProperties)) {
-
-            final int responseCode = connection.getResponseCode();
-
-            /* Fully consuming current request helps with connection re-use
-             * See http://docs.oracle.com/javase/1.5.0/docs/guide/net/http-keepalive.html */
-            HttpUtils.consumeAndCloseConnectionSilently(connection);
-
-            final Map<String, List<String>> header = connection.getHeaderFields();
-            for (final Map.Entry<String, List<String>> entry : header.entrySet()) {
-                LOG.info("Key : {} ,Value : {}", entry.getKey(), entry.getValue());
-            }
-            /*
-             * Do this only on 301,302,303(?)307,308>
-             * Now setting value for all, and lets upper stack to handle it
-             */
-            final String possibleRedirect = connection.getHeaderField("Location");
-
-            final URL redirectUrl;
-            if (possibleRedirect != null && possibleRedirect.trim().length() > 0) {
-                redirectUrl = new URL(possibleRedirect);
-            } else {
-                redirectUrl = null;
-            }
-
-            final long lastModified = connection.getLastModified();
-            final long length = connection.getContentLength();
-
-            return new UrlRequestResult(responseCode, redirectUrl, lastModified, length);
-        }
-
     }
 
     @Override
@@ -128,7 +74,7 @@ public class ResourceDownloader implements Runnable {
 
     private void initializeOnlineResource() {
         try {
-            final UrlRequestResult finalLocation = findBestUrl(resource);
+            final ResourceUrlCreator.UrlRequestResult finalLocation = ResourceUrlCreator.findBestUrl(resource);
             if (finalLocation != null) {
                 initializeFromURL(finalLocation);
             } else {
@@ -144,15 +90,15 @@ public class ResourceDownloader implements Runnable {
         }
     }
 
-    private void initializeFromURL(final UrlRequestResult location) throws IOException {
+    private void initializeFromURL(final ResourceUrlCreator.UrlRequestResult location) throws IOException {
         CacheEntry entry = new CacheEntry(resource.getLocation(), resource.getRequestVersion());
         entry.lock();
-        try (final CloseableConnection connection = getDownloadConnection(location.redirectUrl)) {// this won't change so should be okay not-synchronized
-            resource.setDownloadLocation(location.redirectUrl);
+        try (final CloseableConnection connection = getDownloadConnection(location.getRedirectURL())) {// this won't change so should be okay not-synchronized
+            resource.setDownloadLocation(location.getRedirectURL());
 
             File localFile = CacheUtil.getCacheFile(resource.getLocation(), resource.getDownloadVersion());
-            long size = location.length;
-            long lm = location.lastModified;
+            long size = location.getLength();
+            long lm = location.getLastModified();
             boolean current = CacheUtil.isCurrent(resource.getLocation(), resource.getRequestVersion(), lm) && resource.getUpdatePolicy() != UpdatePolicy.FORCE;
             if (!current) {
                 if (entry.isCached()) {
@@ -253,74 +199,6 @@ public class ResourceDownloader implements Runnable {
 
     }
 
-    /**
-     * Returns the 'best' valid URL for the given resource. This first adjusts
-     * the file name to take into account file versioning and packing, if
-     * possible.
-     *
-     * @param resource the resource
-     * @return the best URL, or null if all failed to resolve
-     */
-    protected UrlRequestResult findBestUrl(final Resource resource) {
-        DownloadOptions options = resource.getDownloadOptions();
-        if (options == null) {
-            options = new DownloadOptions(false, false);
-        }
-
-        List<URL> urls = ResourceUrlCreator.getUrls(resource, options);
-        LOG.debug("Finding best URL for: {} : {}", resource.getLocation(), options.toString());
-        LOG.debug("All possible urls for {} : {}", resource.toString(), urls);
-        for (final HttpMethod requestMethod : validRequestMethods) {
-            for (int i = 0; i < urls.size(); i++) {
-                URL url = urls.get(i);
-                try {
-                    Map<String, String> requestProperties = new HashMap<>();
-                    requestProperties.put(ACCEPT_ENCODING, PACK_200_OR_GZIP);
-
-                    UrlRequestResult response = getUrlResponseCodeWithRedirectionResult(url, requestProperties, requestMethod);
-                    if (response.responseCode == 511) {
-                        if (!InetSecurity511Panel.isSkip()) {
-
-                            boolean result511 = SecurityDialogs.show511Dialogue(resource);
-                            if (!result511) {
-                                throw new RuntimeException("Terminated on users request after encountering 'http 511 authentication'.");
-                            }
-                            //try again, what to do with original resource was nowhere specified
-                            i--;
-                            continue;
-                        }
-                    }
-                    if (response.shouldRedirect()) {
-                        if (response.redirectUrl == null) {
-                            LOG.debug("Although {} got redirect {} code for {} request for {} the target was null. Not following", resource.toString(), response.responseCode, requestMethod, url.toExternalForm());
-                        } else {
-                            LOG.debug("Resource {} got redirect {} code for {} request for {} adding {} to list of possible urls", resource.toString(), response.responseCode, requestMethod, url.toExternalForm(), response.redirectUrl.toExternalForm());
-                            if (!JNLPRuntime.isAllowRedirect()) {
-                                throw new RedirectionException("The resource " + url.toExternalForm() + " is being redirected (" + response.responseCode + ") to " + response.redirectUrl.toExternalForm() + ". This is disabled by default. If you wont to allow it, run javaws with -allowredirect parameter.");
-                            }
-                            urls.add(response.redirectUrl);
-                        }
-                    } else if (response.isInvalid()) {
-                        LOG.debug("For {} the server returned {} code for {} request for {}", resource.toString(), response.responseCode, requestMethod, url.toExternalForm());
-                    } else {
-                        LOG.debug("best url for {} is {} by {}", resource.toString(), url.toString(), requestMethod);
-                        if (response.redirectUrl == null) {
-                            return response.withRedirectUrl(url);
-                        }
-                        return response; /* This is the best URL */
-
-                    }
-                } catch (IOException e) {
-                    // continue to next candidate
-                    LOG.error("While processing " + url.toString() + " by " + requestMethod + " for resource " + resource.toString() + " got " + e + ": ", e);
-                }
-            }
-        }
-
-        /* No valid URL, return null */
-        return null;
-    }
-
     private void downloadResource() {
         URL downloadFrom = resource.getDownloadLocation(); //Where to download from
         URL downloadTo = resource.getLocation(); //Where to download to
@@ -372,7 +250,7 @@ public class ResourceDownloader implements Runnable {
 
     private CloseableConnection getDownloadConnection(URL location) throws IOException {
         final Map<String, String> requestProperties = new HashMap<>();
-        requestProperties.put(ACCEPT_ENCODING, PACK_200_OR_GZIP);
+        requestProperties.put(ResourceUrlCreator.ACCEPT_ENCODING, ResourceUrlCreator.PACK_200_OR_GZIP);
         return ConnectionFactory.openConnection(location, HttpMethod.GET, requestProperties);
     }
 
@@ -466,78 +344,6 @@ public class ResourceDownloader implements Runnable {
 
             outputStream.close();
             inputStream.close();
-        }
-    }
-
-    /**
-     * Complex wrapper around url request Contains return code (default is
-     * HTTP_OK), length and last modified
-     *
-     * The storing of redirect target is quite obvious The storing length and
-     * last modified may be not, but apparently
-     * (http://icedtea.classpath.org/bugzilla/show_bug.cgi?id=2591) the url
-     * connection is not always cached as expected, and so another request may
-     * be sent when length and lastmodified are checked
-     *
-     */
-    static class UrlRequestResult {
-
-        private final int responseCode;
-        private final URL redirectUrl;
-
-        private final long lastModified;
-        private final long length;
-
-        UrlRequestResult(int responseCode, URL redirectUrl, long lastModified, long length) {
-            this.responseCode = responseCode;
-            this.redirectUrl = redirectUrl;
-            this.lastModified = lastModified;
-            this.length = length;
-        }
-
-        UrlRequestResult withRedirectUrl(URL url) {
-            return new UrlRequestResult(responseCode, url, lastModified, length);
-        }
-
-        URL getRedirectURL() {
-            return redirectUrl;
-        }
-
-        int getResponseCode() {
-            return responseCode;
-        }
-
-        /**
-         * @return whether the result code is redirect one. Right now 301-303 and 307-308
-         */
-        boolean shouldRedirect() {
-            return (responseCode == 301
-                    || responseCode == 302
-                    || responseCode == 303 /*?*/
-                    || responseCode == 307
-                    || responseCode == 308);
-        }
-
-        /**
-         * @return whether the return code is not a OK one - anything except <200 or >=300
-         */
-        boolean isInvalid() {
-            return (responseCode < 200 || responseCode >= 300);
-        }
-
-        @Override
-        public String toString() {
-            return ""
-                    + "url: " + (redirectUrl == null ? "null" : redirectUrl.toExternalForm()) + "; "
-                    + "result:" + responseCode + "; "
-                    + "lastModified: " + lastModified + "; "
-                    + "length: " + length + "; ";
-        }
-    }
-
-    private static class RedirectionException extends RuntimeException {
-        RedirectionException(String string) {
-            super(string);
         }
     }
